+++ src/main/java/prismatic/shards/stellarity/key/StellarityRecipes.java
@@ -1,0 +1,7 @@
+package prismatic.shards.stellarity.key;
+
+public interface StellarityRecipes {
+
+
+}
+