+++ src/main/java/prismatic/shards/stellarity/registry/StellarityItems.java
@@ -44,6 +44,7 @@ public interface StellarityItems {
 	Item ROOTED_ENDER_DIRT = registerBlock(StellarityItemIds.ROOTED_ENDER_DIRT, StellarityBlocks.ROOTED_ENDER_DIRT);
 	Item ENDER_DIRT_PATH = registerBlock(StellarityItemIds.ENDER_DIRT_PATH, StellarityBlocks.ENDER_DIRT_PATH);
 	Item ALTAR_OF_THE_ACCURSED = registerBlock(StellarityItemIds.ALTAR_OF_THE_ACCURSED, StellarityBlocks.ALTAR_OF_THE_ACCURSED);
+	Item ENDERITE_BLOCK = registerBlock(StellarityItemIds.ENDERITE_BLOCK, StellarityBlocks.ENDERITE_BLOCK);
 
 	Item CALL_OF_THE_VOID = register(StellarityItemIds.CALL_OF_THE_VOID, CallOfTheVoid::new, CallOfTheVoid.PROPERTIES);
 	Item FISHER_OF_VOIDS = register(StellarityItemIds.FISHER_OF_VOIDS, FisherOfVoids::new, FisherOfVoids.PROPERTIES);
@@ -201,7 +202,7 @@ public interface StellarityItems {
 	Item VOIDED_SILVERFISH_SPAWN_EGG = registerSpawnEgg(StellarityItemIds.VOIDED_SILVERFISH_SPAWN_EGG, StellarityEntities.VOIDED_SILVERFISH);
 	Item VOIDED_SLIME_SPAWN_EGG = registerSpawnEgg(StellarityItemIds.VOIDED_SLIME_SPAWN_EGG, StellarityEntities.VOIDED_SLIME);
 	Item POTASSIFISH = register(StellarityItemIds.POTASSIFISH, foodProperties(1, 0.6f, new MobEffectInstance(MobEffects.REGENERATION, 4 * 20)));
-	
+
 	static Supplier<ItemStack> createPotion(Holder<Potion> potion) {
 		return () -> PotionContents.createItemStack(Items.POTION, potion);
 	}